refactor(7084): isolate verified actor evidence policy
The Agent C slice now keeps the shared evidence shape separate, restores the coordination module below the repository size ceiling, and rejects output paths that alias source artifacts.

Constraint: Every source file must remain at or below 200 lines
Rejected: Rely on existing-file conflicts for source path aliases | it obscures the artifact boundary and produces the wrong failure
Confidence: high
Scope-risk: narrow
Directive: Keep source validation shared and Agent C disclosure policy isolated
Tested: 29 Pi extension tests; maximum touched TypeScript file is 198 lines; diff check clean
Not-tested: Live three-process integration rehearsal
